Clear chart before unmount
Google Chart creates a lot of nodes in the body for elements
in the legend, that should be truncated with ellipsis.
Later these nodes will be used as a tooltip when user hover over
a truncated element.
When component is unmounted, these orphan nodes are left in the body,
which leads to memory leak and a worse DOM performance because of
a lot of elements in it. This is especially true in SPAs, when you might
have one page with google charts which might affect the rest of the
pages if you don't remove orphan nodes manually.

To test this change you can check that there is a single node, created
in the body for "Very long label...".
When you click "Toggle orphans demo", this node will be removed.